numerology_api/pyproject.toml
[project]
name = "numerology-api"
version = "1.0.0"
description = "FastAPI web service for numerology calculations"
requires-python = ">=3.10.17"
dependencies = [
    "numerology==1.0.0",
    "numerology-cli==1.0.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pyjwt>=2.9.0",
    "bcrypt>=4.0.0",
    "httpx>=0.27.0",
    "python-multipart>=0.0.12",
    "sqlmodel>=0.0.22",
    "python-dotenv>=1.0.1",
]

[project.scripts]
numerology-api = "numerology_api.main:start"

[tool.uv.sources]
numerology = { path = "../numerology", editable = true }
numerology-cli = { path = "../numerology_cli", editable = true }

[[tool.uv.index]]
url = "https://mirrors.aliyun.com/pypi/simple"
default = true

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
